Deal with nested clusters in the clusterplot
Previously, larger clusters would 'overwrite' smaller ones that they
contain. Now the smaller ones are kept as separate clusters. Each sequence
can still be only part of a single cluster, so the larger cluster will not
contain the sequences of nested clusters. This may be a bug or exactly what
we want.