Create a much better test data
Four genes are actually expressed in this data set at such a level that the
filters are passed. All the plots are therefore created (dendrogram, error
frequencies, clusterplots), and the pipeline finishes its two iterations
without error. It doesn’t really discover anything, but that can be done
later. Something like this should be included with the final version so that
people can check whether their installation works.