Fix IgBLAST not aligning first base
IgBLAST sometimes starts the alignment at the second base of the subject.
For example, if the reference is CAGGTGCAG, and the query contains GAGGTACAG,
then IgBLAST ignores the first base and aligns only AGGTGCAG to AGGTACAG.
This leads the discovery part of the pipeline to ignore those sequences
since they do not start with one of the required 5' motifs.