Enable doctests
This also made it necessary to add a few matplotlib.use('pdf') lines
to the subcommnads. They are all imported one after the other during
the tests and if that line is omitted, matplotlib tries to import
PyQt or PySide, which fails.

matplotlib should really support some way to use it without setting
global state like that.