Make it possible to whitelist known sequences
If a found gene candidate appears in that list, the sequence is included
in the list of discovered sequences even when it would otherwise not
pass filtering criteria. To enable this, add a ``whitelist.fasta`` file
to the project directory before starting the analysis.

This feature has not yet been integrated into the 'init' subcommand.