Fix whitelisting bug
To make it clearer what is happening, avoid the term 'merging' in the
description of the merged() method: It determines which of the sequences to
discard (if any).

Previously, both sequences would be kept if any was whitelisted, no matter
what.

The intended behavior, implemented now, is to prevent a sequence from being
discarded if it is whitelisted.