Use dynamic threshold for plotting error histograms
Until now, at least 200 sequences were required for a gene’s error histogram
to be plotted. This change introduces a dynamic threshold of 0.1% of the
number of assigned sequences, but at most 100. With normal-sized datasets
(~1 million reads), the 100 ends up being used, but with smaller datasets,
the threshold is lowered automatically.