Decrease default threshold for merging candidates to 1
When we merge candidates that are two differences apart, we see in some
experiments that we miss true novel genes.

It may also be that the reason we initially needed this was because of the
cross-mapping problem, but at the time, we did not know about it. Since we
deal with cross-mapping problems in a much better way now, there may be no
reason to keep this differences setting at all, but that requires further
experiments. Lowering it to 1 should be uncontroversial.