Be stricter when re-adding "undiscovered" db sequences
Previously, the database sequence was counted as expressed if there were
sequences with V_SHM <= 1%. Also, the cluster_size column reported the number
of sequences with V_SHM <= 1%. This is too broad.

Now, the database sequence is seen as expressed if there are some (5) exact
occurrences of it. Only then will it be added to the list of candidates even
if it was not found as a cluster. The cluster_size column for the "db" genes
is now identical to the "exact" column (and therefore much smaller).

Closes #48