Add a configuration setting that can turn off consensus taking of barcode groups
When barcoding is enabled, sequences that have identical barcode and CDR3 are
usually collapsed into a single consensus sequence. This change allows one to
set barcode_consensus: false in the configuration file, and then no
collapsing and consensus taking is done and only the barcode is removed from
each sequence.

Closes #68