Speed up grouping by barcode
The speedup is extreme when using --pseudo-cdr3: Processing time goes down
from hours to minutes. When using the real CDR3, it is ”only” twice as fast.

The results change very slightly since the clustering algorithm works in a
slightly different way, but it does not seem to be better or worse.