Add column "missing" to discoverj output
A candidate D gene is now given the name of the database gene even if
nucleotides are missing in the beginnig or in the end. So if the 'name'
column contains IGHD6-6*01, then it is possible that the discovered sequence
is actually not quite the one of IGHD6-6*01. To make this clear, a column
named 'missing' was added, which contains the missing nucleotides.

If the value is AC...GT in, for example, then this means that, in order to
actually get to the full database sequence, one would need to add AC in the
front and GT at the end (starting from the sequence given in the 'sequence'
column).