clonoquery: Ignore different lengths when comparing CDR3s
This tries to fix an issue where the query and reference table use two
different versions of IgDiscover to do CDR3 assignments.

This issue occurs only in rare cases when the length of the nucleotide
sequence of a CDR3 is not divisible by three in both query and reference
tables.