discover: Add --cluster-exact option
This can be used to generate candidates from multiple occurrences of
identical V sequences. Interestingly, it also helps to find longer
candidates. In SRR2905698, it helps to find the IGHV1-2*02_S4497
candidate rather than the truncated version IGHV1-2*02_S4820.

One advantage of this method is that it could work without V assignments
although it needs to know where the V gene is located (start and end
within the read). Since no consensus sequences need to be computed, it
is also fast.