discover: Group by J before counting CDR3s
When computing cdr3_clusters for a candidate that has a large number of
exact matches in the dataset, the cdr3_clusters number can decrease because
pure chance makes it more likely that clusters become connected to each
other.

To counter this somewhat, we now first group by J gene and then cluster
by CDR3 within each group. The cdr3_clusters number is then the
sum of the number of clusters in each group.