haplotype: When het gene turns out to be non-het, try a different one
When the V gene that was chosen to haplotype the Js is not actually
heterozygous, but duplicated, then the haplotyping for the Js will be wrong
(they will also be called duplicate).

We cannot prevent this from happening since when the program chooses a
heterozygous J, it does not know whether it will later on actually classify
it as 'duplicate' or perhaps 'unknown'.

What the program can do and now does is to detect the situation: If the most
highly expressed assumed to be heterozygous V or J turns out not to be
heterozygous, it proceeds with the second-most highly expressed ones and so
on. It attempts up to five heterozygous genes. If after five attempts the
described situation is still present, it gives up and just uses the first
solution that it computed (the one using the most highly expressed het.
V/J).

This should fix issue #101 for some datasets. SRR5408023 is an example.