discoverj: Reduce hard-coded min_occ to 3 (from 100)
This means that a candidate is generated if there are at least 3 exact
occurrences of that sequence, not 100.

This helps to recover more alleles in datasets with low numbers of reads,
and also recovers more full-length sequences. It does not lead to more
false positives in datasets with high number of reads.