discoverj: Auto-whitelist all J alleles that are expressed
Expressed Js are computed in the same way as in 'igdiscover count'. That is,
all alleles in the expressed_J.tab are whitelisted.

This change does not actually do anything on my current test datasets as the
previous optimizations (setting min-count=1 in particular) made the missing
alleles re-appear already. Whitelisting may possibly still be necessary on
really bad datasets.