Test whether specifying primer sequences leads to a SyntaxError
Snakemake version 5.0 and newer complain about the trim_forward_primers
rule with this error message:

SyntaxError:
Not all output, log and benchmark files of rule trim_forward_primers contain the same
wildcards. This is crucial though, in order to avoid that two or more jobs write to the
same file.

The error is only triggered if forward_primers are defined in the
configuration.

See issue #93