germlinefilter: Discard only identical sequences, not similar ones
This removes the TooSimilarSequenceFilter and replaces it with an
IdenticalSequenceFilter that does less. Previously, two sequences would be
considered to be “too similar” if they are identical, ignoring those
nucleotides in the 3’ end that are the beginning of the CDR3. This is too
sensitive as it would also remove legit novel alleles that happen to differ
from a known one in those nucleotides of the 3’ end. For light chain V
sequences, this was even more problematic as the CDR3 covers more nucleotides
within the V.

When deciding whether a candidate should be removed, the new filter uses
these simpler criteria:

• If the target candidate is whitelisted, do nothing.
• If the target candidate is a truncated version of another one, but
  otherwise identical, remove it.
• If the target candidate is identical to another one, remove it.

Comparisons use the *full length* V sequences.

With this change, we recover at least one more allele (IGHV3-43_S5024) in
three of the regression testing datasets.