igblast: Fix name-mangling issue
If IgBLAST (or BLAST) recognizes a sequence id as being an accession, it
mangles the name to conform to its naming scheme (GQ123456 becomes
"gb|GQ123456|") such that our parsing routines don’t recognize the name
anymore. To avoid this, we now mangle names ourselves by prefixing them with
'%'. This is only done in the temporary BLAST database and undone during
parsing, so it should be transparent to the user.

Added also some handling of sequence ids with whitespace.

Close #100