Specify abstract dependencies in environment.yml
Add an environment.lock.yml file with "locked" dependencies and a script
misc/condalock.sh that creates it from environment.yml.

The reason for simplifying this is that packages disappear from anaconda.org
quite often and then the locked environment needs to be updated.

Also, there is now a weekly Travis test that will (implicitly) check whether
the environment.lock.yml file is still working.