Decrease required cluster size to 50
Since switching to IgBLAST 1.10, default settings were missing the verified
novel allele IGHV3-13_S3164 in the ERR1760498 dataset because its cluster
size dropped to 99.  Even the other novel allele IGHV3-20_S2478 only has a
cluster size of 101, so it was included due to chance.  To reduce the chance
of this happening in this and similar cases, this commit decreases the
required cluster size to 50.

Additionally, also IGHV4-28_S5306 will now be found.  It is not novel (it is
IGHV4-28*01 in the IMGT database), but is missing from the database we use
for testing.