Add augment subcommand
This takes in an AIRR-formatted table and adds IgDiscover-specific columns
to it. Also, the CDR3 sequence is detected.

The output table contains both the AIRR columns and the IgDiscover-specific
columns. Many of them are duplicated, but in this way, it should be possible
to use this table as input wherever any of the other IgDiscover subcommands
expects an ".assigned.tab" file.